Convert `AuthorizationModal` to TSX
